feat(deps)!: Update tree-sitter bindings to v0.23.0
Update to tree-sitter v0.23.0.  The change contains breaking changes,
which are explained in the links below.

What the update does is: it removes the dependency on tree-sitter from
the bindings.  Now clients can chose their own tree-sitter version to
use the bindings with.

Most changes are form the auto-updated bindings.

This commit also fixes the Go build.

See
- https://github.com/alex-pinkus/tree-sitter-swift/pull/435
- https://github.com/tree-sitter/tree-sitter/pull/3069